Create spellings.txt
Strings in the English.pak file that need to be changed.

This is a list, not actual code. Still, this will form the basis for a script to make the appropriate substitutions.

Please raise an issue or make a pull request if there are American spellings absent from this list.